+++ src/constants.js
@@ -21,7 +21,7 @@ export const EXIT_CODES = {
 /**
  * These events are just internal for Apify package, so we don't need them in apify-shared package.
  *
- * @type {{CPU_INFO: string, SYSTEM_INFO: string, MIGRATING: string, PERSIST_STATE: string}}
+ * @type {{CPU_INFO: string, SYSTEM_INFO: string, MIGRATING: string, PERSIST_STATE: string, ABORTING: string}}
  */
 export const ACTOR_EVENT_NAMES_EX = { ...ACTOR_EVENT_NAMES, PERSIST_STATE: 'persistState' };