+++ test/crawlers/cheerio_crawler.test.js
@@ -723,7 +723,7 @@ describe('CheerioCrawler', () => {
         });
 
         test('should markBad sessions after request timeout', async () => {
-            log.setLevel(log.LEVELS.OFF);
+            // log.setLevel(log.LEVELS.OFF);
             const cheerioCrawler = new Apify.CheerioCrawler({
                 requestList: await Apify.openRequestList(null, [
                     `http://${HOST}:${port}/timeout?a=12`,
@@ -733,7 +733,9 @@ describe('CheerioCrawler', () => {
                 requestTimeoutSecs: 1,
                 maxConcurrency: 1,
                 useSessionPool: true,
-                handlePageFunction: async () => {},
+                handlePageFunction: async () => {
+                    await Apify.utils.sleep(1);
+                },
             });
 
             await cheerioCrawler.run();
@@ -750,7 +752,7 @@ describe('CheerioCrawler', () => {
                 expect(session.errorScore).toEqual(1);
             });
 
-            log.setLevel(log.LEVELS.ERROR);
+            // log.setLevel(log.LEVELS.ERROR);
         });
 
         test('should retire session on "blocked" status codes', async () => {